+++ website/flake.nix
@@ -46,17 +46,18 @@
         in
         {
           apps = {
-            dev = mkApp "dev" [ pkgs.nodejs ] "npm run dev";
-            build = mkApp "build" [ pkgs.nodejs ] "npm run build";
-            preview = mkApp "preview" [ pkgs.nodejs ] "npm run preview";
+            dev = mkApp "dev" [ pkgs.nodejs pkgs.pnpm ] "pnpm run dev";
+            build = mkApp "build" [ pkgs.nodejs pkgs.pnpm ] "pnpm run build";
+            preview = mkApp "preview" [ pkgs.nodejs pkgs.pnpm ] "pnpm run preview";
             deploy =
               mkApp "deploy"
                 [
                   pkgs.nodejs
+                  pkgs.pnpm
                   pkgs.firebase-tools
                 ]
                 ''
-                  npm run build
+                  pnpm run build
                   firebase deploy --only hosting
                 '';
           };
@@ -64,6 +65,7 @@
           devShells.default = pkgs.mkShellNoCC {
             packages = [
               pkgs.nodejs
+              pkgs.pnpm
               pkgs.firebase-tools
             ];
           };